move models to server dir
