+++ server/middlewares.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/Daskott/kronus/colors"
-	"github.com/Daskott/kronus/models"
+	"github.com/Daskott/kronus/server/models"
 	"github.com/gorilla/mux"
 )
