+++ server/work/reaper.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/Daskott/kronus/colors"
-	"github.com/Daskott/kronus/models"
+	"github.com/Daskott/kronus/server/models"
 	"gorm.io/gorm"
 )
